[!!!][FEATURE] Introduce upgrade:run and upgrade:prepare
The command upgrade:all and upgrade:wizard are removed.
They are superseded by upgrade:prepare and upgrade:run.

Instead of implicitly performing upgrade preparation steps
when calling upgrade:list and upgrade:all, we now introduce
a dedicated command upgrade:prepare. Preparation steps
and extension compatibility checks are performed on each
invocation of this command. Having executed this command once
is a prerequisite of calling upgrade:list or upgrade:run

upgrade:run can take an argument, which can be a wizard identifier
of or the keyword "all". "upgrade:run all" will run
all wizards that are due for execution.

Fixes: #859
Fixes: #859
Fixes: #805